Price Cleaning using Google Sheets
This is one of my Data Cleaning Project using Google Sheets. The Goal is to analyze whether the price (in IDR Rupiah Indonesian) is suitable for each Video Calls, Audio Calls, and Chat method. The data is taken from a survey, and I just used a part of the survey which consist of the pricing preferences.